✨ Setup POST/PUT/DELETE endpoints without response
