+++ src/sentry/api/endpoints/notifications/notification_actions_details.py
@@ -1,8 +1,10 @@
+from rest_framework import status
 from rest_framework.response import Response
 
 from sentry.api.base import region_silo_endpoint
 from sentry.api.bases.organization import OrganizationEndpoint
 from sentry.api.serializers import Serializer, serialize
+from sentry.api.serializers.rest_framework.notification_action import NotificationActionSerializer
 from sentry.models.notificationaction import NotificationAction
 from sentry.models.organization import Organization
 
@@ -26,13 +28,27 @@ class NotificationActionsDetailsEndpoint(OrganizationEndpoint):
 
     def put(self, request, organization: Organization, action_id: int):
         try:
-            NotificationAction.objects.get(
+            action = NotificationAction.objects.get(
                 id=action_id,
                 organization_id=organization.id,
             )
         except NotificationAction.DoesNotExist:
             return Response(status=404)
-        return Response(status=201)
+
+        serializer = NotificationActionSerializer(
+            instance=action,
+            context={
+                "access": request.access,
+                "organization": organization,
+                "user": request.user,
+            },
+            data=request.data,
+        )
+        if not serializer.is_valid():
+            return Response(serializer.errors, status=400)
+
+        serializer.save()
+        return Response(status=status.HTTP_202_ACCEPTED)
 
     def delete(self, request, organization: Organization, action_id: int):
         try:
@@ -43,4 +59,4 @@ class NotificationActionsDetailsEndpoint(OrganizationEndpoint):
         except NotificationAction.DoesNotExist:
             return Response(status=404)
         action.delete()
-        return Response(status=204)
+        return Response(status=status.HTTP_204_NO_CONTENT)
